New buffer API (and support for reading extra bytes in LAS files) (#22)
* Added parsing of some known LAS VLRs

* First draft of new parser architecture

* Tests for parsing of LAS attributes with conversions

* Fixed some tests due to change in how Vector3 conversions work (affects COLOR_RGB attribute)

* First tests with extra bytes data. Added helper binary to generate test data with extra bytes

* Caching known parsers when reading into custom layout multiple times. More tests.

* Massive refactoring of PointBuffer API, WIP

* Added converting attribute view. Refactored pasture-algorithms to use new point buffer API

* Removed 'bool' type and added 'Vec3i32' type to pasture-derive

* WIP refactoring of pasture-io to new buffer API

* Refactoring 3D Tiles reader and writer for new buffer API

* Refactored ASCII reader and writer for new buffer API

* Refactored pasture-tools for new buffer API

* Documentation

* Fixed doctests. Fixed bug for buffer created from empty point layout

* Fixed bug that incorrectly assumed that a VLR was of a known type. Fixed bug when reading LAS files that have leftover bytes between the end of VLRs and the beginning of the point records

* Small performance improvements to PointLayout and PointAttributeDefinition

* Added sort function to PointViewMut

* Added example for mmaped LAS parsing

* applied clippy suggestions to pasture-core

* applied clippy suggestions to pasture-algorithms

* applied clippy suggestions to pasture-derive

* Applied clippy suggestions to pasture-io

* applied clippy suggestions to pasture-tools

* Added BufferLayoutConverter to convert buffers into different PointLayouts

* Debugging LAS I/O and adding more tests

* Upgrade las-rs and laz-rs crates to support some more point formats in readers/writers

* Tests for LAZ readers with formats 4 and 5. Higher formats still unsupported unfortunately

* More documentation. Added an example that illustrates buffer layout conversion

* Add FromIterator implementation for PointLayout

* Fixed bug with unused VLRs in LAZ reader that caused the byte offset to point data in the LAS header to be wrong

* Refactoring LAS/LAZ reading to use BufferLayoutConverter. Changed PointReader API to have more relaxed requirements for buffer types

* Added raw attribute views to improve performance of buffer layout conversion for interleaved buffers

* Benchmark for layout conversions

* Convenience function to construct AABB from iterator of positions

* O(1) nth implementation for point and attribute iterators

* Added filter function for HashMapBuffer

* Read LAS custom layout in chunks for memory efficiency

* Removed old GPU code

* Bump version and updated readme

---------

+++ pasture-io/src/tiles3d/common.rs
@@ -46,7 +46,7 @@ pub fn write_json_header<W: Write>(
 
     if num_padding_bytes > 0 {
         writer
-            .write(&vec![0x20; num_padding_bytes as usize])
+            .write(&vec![0x20; num_padding_bytes])
             .context("Could not write padding bytes of JSON header to writer")?;
     }
